Complete docsdocstrings
